requirements.txt
django
alembic
maybe-else
numpy
pandas
pathmagic
pymiscutils
pysubtypes
sqlalchemy
sqlparse
pytest
tabulate
